refactor(review): extract wire envelopes to application/review_envelope.rs (E2)
ObjectDiffEnvelope and ReviewEnvelope are serialization-shaped DTOs
the CLI and MCP layers turn into JSON. They were stranded inside
application/review.rs (1142 lines after E1) where the snapshot →
compile → diff → enrich orchestration also lives. Move them into a
sibling application/review_envelope.rs so review.rs stays orchestration-
focused.

Visibility remains pub for now — adoc-local, adoc-cli, adoc-mcp still
import the types. E3 demotes them to pub(crate) once lib.rs exposes
build_*_envelope_value(...) -> serde_json::Value as the public API.

application/review.rs: 1142 → ~770 lines. Wire output is unchanged
(types are byte-identical, just relocated); cargo test --workspace
green; contract schemas for adoc.diff.v0 and adoc.review.v0 still
validate byte-equally.
